memory-only

+++ flex/storages/rt_mutable_graph/mutable_property_fragment.cc
@@ -109,7 +109,8 @@ inline DualCsrBase* create_csr(EdgeStrategy oes, EdgeStrategy ies,
   return nullptr;
 }
 
-void MutablePropertyFragment::Open(const std::string& work_dir) {
+void MutablePropertyFragment::Open(const std::string& work_dir,
+                                   bool memory_only) {
   std::string schema_file = schema_path(work_dir);
   std::string snapshot_dir{};
   bool build_empty_graph = false;
@@ -131,24 +132,39 @@ void MutablePropertyFragment::Open(const std::string& work_dir) {
 
   vertex_data_.resize(vertex_label_num_);
   std::string tmp_dir_path = tmp_dir(work_dir);
-  if (std::filesystem::exists(tmp_dir_path)) {
-    std::filesystem::remove_all(tmp_dir_path);
-  }
+  if (!memory_only) {
+    if (std::filesystem::exists(tmp_dir_path)) {
+      std::filesystem::remove_all(tmp_dir_path);
+    }
 
-  std::filesystem::create_directories(tmp_dir_path);
+    std::filesystem::create_directories(tmp_dir_path);
+  }
 
   std::vector<size_t> vertex_capacities(vertex_label_num_, 0);
   for (size_t i = 0; i < vertex_label_num_; ++i) {
     std::string v_label_name = schema_.get_vertex_label_name(i);
 
-    lf_indexers_[i].open(vertex_map_prefix(v_label_name), snapshot_dir,
-                         tmp_dir_path);
+    if (memory_only) {
+      lf_indexers_[i].open_in_memory(snapshot_dir + "/" +
+                                     vertex_map_prefix(v_label_name));
+    } else {
+      lf_indexers_[i].open(vertex_map_prefix(v_label_name), snapshot_dir,
+                           tmp_dir_path);
+    }
 
-    vertex_data_[i].open(vertex_table_prefix(v_label_name), snapshot_dir,
-                         tmp_dir_path, schema_.get_vertex_property_names(i),
-                         schema_.get_vertex_properties(i),
-                         schema_.get_vertex_storage_strategies(v_label_name));
-    if (!build_empty_graph) {
+    if (memory_only) {
+      vertex_data_[i].open_in_memory(
+          vertex_table_prefix(v_label_name), snapshot_dir,
+          schema_.get_vertex_property_names(i),
+          schema_.get_vertex_properties(i),
+          schema_.get_vertex_storage_strategies(v_label_name));
+    } else {
+      vertex_data_[i].open(vertex_table_prefix(v_label_name), snapshot_dir,
+                           tmp_dir_path, schema_.get_vertex_property_names(i),
+                           schema_.get_vertex_properties(i),
+                           schema_.get_vertex_storage_strategies(v_label_name));
+    }
+    if (!build_empty_graph && !memory_only) {
       vertex_data_[i].copy_to_tmp(vertex_table_prefix(v_label_name),
                                   snapshot_dir, tmp_dir_path);
     }
@@ -195,13 +211,19 @@ void MutablePropertyFragment::Open(const std::string& work_dir) {
             create_csr(oe_strategy, ie_strategy, properties);
         ie_[index] = dual_csr_list_[index]->GetInCsr();
         oe_[index] = dual_csr_list_[index]->GetOutCsr();
-        dual_csr_list_[index]->Open(
-            oe_prefix(src_label, dst_label, edge_label),
-            ie_prefix(src_label, dst_label, edge_label),
-            edata_prefix(src_label, dst_label, edge_label), snapshot_dir,
-            tmp_dir_path);
+        if (memory_only) {
+          dual_csr_list_[index]->OpenInMemory(
+              oe_prefix(src_label, dst_label, edge_label),
+              ie_prefix(src_label, dst_label, edge_label),
+              edata_prefix(src_label, dst_label, edge_label), snapshot_dir);
+        } else {
+          dual_csr_list_[index]->Open(
+              oe_prefix(src_label, dst_label, edge_label),
+              ie_prefix(src_label, dst_label, edge_label),
+              edata_prefix(src_label, dst_label, edge_label), snapshot_dir,
+              tmp_dir_path);
+        }
         ie_[index]->resize(vertex_capacities[dst_label_i]);
-
         oe_[index]->resize(vertex_capacities[src_label_i]);
       }
     }
